+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+# 0.0.6
+- It cares about height now.
+- You can use:
+```js
+DHDistances.Toggle({mode: '2d'}); // Calculates distance ignoring elevation.
+```
+```js
+DHDistances.Toggle({mode: '3d'}); // Calculates distance accounting for elevation.
+```
+```js
+DHDistances.Toggle({mode: 'both'}); // Shows both 3D and 2D distances.
+```
+
 # 0.0.5
 - light animation